chatbot/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
tensorflow>=2.16.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
